refactor: drop litegraph uuid shim, point importers at @/utils/uuid
Every consumer was already going through a 2-line re-export at
src/lib/litegraph/src/utils/uuid.ts. Rewrites the 20 importers (10 in
src/lib/litegraph/, 10 outside) to import from @/utils/uuid directly
and deletes the shim. The litegraph.ts barrel still re-exports
createUuidv4/UUID for the package's public surface.

Amp-Thread-ID: https://ampcode.com/threads/T-019e71ba-da65-71c1-b43d-bab1eb1c7312
