+++ src/lib/litegraph/src/subgraph/EmptySubgraphOutput.ts
@@ -3,7 +3,7 @@ import type { LLink } from '@/lib/litegraph/src/LLink'
 import type { RerouteId } from '@/lib/litegraph/src/Reroute'
 import type { INodeOutputSlot, Point } from '@/lib/litegraph/src/interfaces'
 import { nextUniqueName } from '@/lib/litegraph/src/strings'
-import { zeroUuid } from '@/lib/litegraph/src/utils/uuid'
+import { zeroUuid } from '@/utils/uuid'
 
 import { SubgraphOutput } from './SubgraphOutput'
 import type { SubgraphOutputNode } from './SubgraphOutputNode'
