--- src/lib/litegraph/src/utils/uuid.ts
@@ -1,2 +1,0 @@
-export type { UUID } from '@/utils/uuid'
-export { createUuidv4, zeroUuid } from '@/utils/uuid'
