+++ src/stores/widgetValueStore.ts
@@ -1,7 +1,7 @@
 import { defineStore } from 'pinia'
 
 import type { NodeId } from '@/lib/litegraph/src/LGraphNode'
-import type { UUID } from '@/lib/litegraph/src/utils/uuid'
+import type { UUID } from '@/utils/uuid'
 import type { ComponentKey } from '@/world/componentKey'
 import {
   asGraphId,
